organized and clean up code
