+++ app/api/auth/[...nextauth]/route.tsx
@@ -1,5 +1,5 @@
 import NextAuth, { NextAuthOptions } from "next-auth";
-import prisma from "../../../libs/prismadb";
+import prisma from "@/app/libs/prismadb";
 import { PrismaAdapter } from "@next-auth/prisma-adapter";
 import CredentialsProvider from "next-auth/providers/credentials";
 import GoogleProvider from "next-auth/providers/google";
@@ -98,36 +98,35 @@ export const authOptions: NextAuthOptions = {
   },
   callbacks: {
     async signIn({ account, profile }) {
-		if (account?.provider === "google" && profile) {
-			const existingUser = await prisma.user.findUnique({
-			  where: {
-				email: profile?.email!,
-				NOT: {
-				  provider: "google",
-				},
-			  },
-			});
-			if (existingUser) {
-			  return false;
-			}
-		  }
-	
-		  if (account?.provider === "facebook" && profile) {
-			const existingUser = await prisma.user.findUnique({
-			  where: {
-				email: profile?.email!,
-				NOT: {
-				  provider: "facebook",
-				},
-			  },
-			});
-			if (existingUser) {
-			  return false;
-			}
-		  }
-	
-		  return true;
-		
+      if (account?.provider === "google" && profile) {
+        const existingUser = await prisma.user.findUnique({
+          where: {
+            email: profile?.email!,
+            NOT: {
+              provider: "google",
+            },
+          },
+        });
+        if (existingUser) {
+          return false;
+        }
+      }
+
+      if (account?.provider === "facebook" && profile) {
+        const existingUser = await prisma.user.findUnique({
+          where: {
+            email: profile?.email!,
+            NOT: {
+              provider: "facebook",
+            },
+          },
+        });
+        if (existingUser) {
+          return false;
+        }
+      }
+
+      return true;
     },
     async jwt({ token, user, session, account, profile }) {
       console.log("jwt callback", { token, user, session, account, profile });
