+++ app/api/user/profile/[id]/route.tsx
@@ -1,4 +1,4 @@
-import prisma from "../../../../libs/prismadb";
+import prisma from "@/app/libs/prismadb";
 import { NextResponse } from "next/server";
 
 export async function GET(request: Request, { params }: { params: { id: string } }) {
