+++ app/libs/interfaces.tsx
@@ -30,3 +30,50 @@ export interface RequestProps {
   requesterImage: string;
   requesterCity: string;
 }
+
+export interface ImageMapping {
+  [key: string]: string;
+}
+
+export interface APIErr {
+  code: number;
+  message: string;
+  cause: string | Error;
+}
+
+export interface CardProps {
+  smallCard?: boolean;
+  request?: RequestProps;
+}
+
+// Define interfaces for location data
+export interface LocationFeature {
+  geometry: {
+    coordinates: [number, number];
+  };
+  place_name: string;
+  context: Array<{ text: string }>;
+}
+
+export interface LocationData {
+  features: LocationFeature[];
+}
+
+// Define an interface for your form data
+export interface FormData {
+  name: string;
+  birthday: string;
+  ethnicity: string;
+  gender: string;
+  phonenumber: string;
+  location: {
+    lng: number;
+    lat: number;
+    address: {
+      fullAddress: string;
+      pointOfInterest: string;
+      city: string;
+      country: string;
+    };
+  };
+}
